number bucket

+++ programs/renec-dapp-template/src/constants.rs
@@ -1,6 +1,6 @@
 use anchor_lang::prelude::Pubkey;
 
-pub const DISCRIMINATOR_SIZE: usize = std::mem::size_of::<u8>();
+pub const DISCRIMINATOR_SIZE: usize = std::mem::size_of::<u64>();
 pub const PUBKEY_SIZE: usize = std::mem::size_of::<Pubkey>();
 pub const U8_SIZE: usize = std::mem::size_of::<u8>();
 pub const U32_SIZE: usize = std::mem::size_of::<u32>();
